Add files via upload
Before, all the algorithms were used to sort arrays with only 1000 values. In this newer version of the program, I created a global variable that allows the user to change the number of values that the sorting algorithms need to sort. Now, if the user wants all the sorting algorithms to sort 500 values instead of 1000, he simply needs to change the value of the global variable. This also improves the organization of the code since now one variable is used throughout the whole code instead of multiple different numbers.